Added a frame tolerance and sliders to control it.
An issue I was having with the previous implementation was if a pixel changed color for 2 or 3 frames, the mask would turn off for those frames immediately. The opposite was true as well, If one of the tracked pixels for a specific mask just happened to be the right color for a frame, the mask would activate. 
This update allows for you to control how many frames a pixel has to remain it's specific color to turn on it's mask and how many frames a pixel color has to be wrong to turn off it's mask.